pyproject.toml
[project]
name = "deathroll"
version = "0.0.1"
description = "DeathRoll — Discord-hosted casino platform settling in WoW Gold"
authors = [{ name = "Aleix" }]
requires-python = ">=3.12"
readme = "README.md"
license = { file = "LICENSE" }

dependencies = [
    "discord.py==2.4.0",
    "asyncpg==0.30.0",
    "sqlalchemy==2.0.36",
    "alembic==1.14.0",
    "psycopg2-binary==2.9.10",
    "pydantic==2.10.4",
    "pydantic-settings==2.7.0",
    "structlog==24.4.0",
    # Pillow was added for Luck's banner generator which is paused;
    # no current import touches it, so we drop the dep to remove the
    # CVE surface (GHSA-cfh3-3jmp-rvhc, GHSA-whj4-6x5x-4v2j on
    # 11.0.0). Re-add when Luck resumes — bump to 12.x then.
    "prometheus-client==0.21.0",
]

[dependency-groups]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.25.0",
    "pytest-cov==6.0.0",
    "hypothesis==6.122.3",
    "testcontainers[postgres]==4.9.0",
    "pip-audit==2.7.3",
    "ruff==0.8.4",
    "mypy==1.13.0",
    # types-Pillow dropped alongside Pillow (Story 15.3 security review).
    # Story 11.3: alert rules YAML validation in tests/unit/dw/test_alert_rules.py.
    "pyyaml==6.0.2",
    "types-PyYAML==6.0.12.20240917",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["deathroll_core", "deathroll_luck", "deathroll_deposit_withdraw", "deathroll_poker", "tests"]

[tool.ruff.lint]
select = [
    "E", "F", "W",   # pycodestyle, pyflakes
    "I",             # isort
    "B",             # bugbear
    "UP",            # pyupgrade
    "N",             # pep8-naming
    "S",             # bandit (security)
    "T20",           # print() usage
    "RUF",           # ruff-specific
]
ignore = [
    "S101",          # assert in tests is fine
    "E501",          # line length handled by formatter
]
[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S105", "S106"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
warn_return_any = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false

# asyncpg ships without a py.typed marker; treat it as untyped so the
# bot code can ``import asyncpg`` cleanly under strict mode.
[[tool.mypy.overrides]]
module = ["asyncpg", "asyncpg.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers --strict-config"
markers = [
    "unit: unit tests",
    "integration: integration tests requiring postgres",
    "property: hypothesis-driven property tests",
    "e2e: end-to-end tests against a Discord sandbox",
]

[tool.coverage.run]
source = ["deathroll_core", "deathroll_luck", "deathroll_deposit_withdraw"]
omit = ["**/__init__.py", "tests/**"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
fail_under = 0   # global gate is enforced separately in CI per module
